pyproject.toml
[tool.poetry]
name = "scrappingchef"
version = "0.1.0"
description = ""
authors = ["JeanArnaudRitouret <86613080+JeanArnaudRitouret@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.13.1"
annotated-types = "0.7.0"
asgiref = "3.7.2"
attrs = "24.1.0"
beautifulsoup4 = "4.9.1"
bleach = "6.1.0"
Brotli = "1.1.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
Cython = "3.0.11"
Django = "4.2.3"
django-cors-headers = "4.5.0"
djangorestframework = "3.14.0"
h11 = "0.14.0"
idna = "3.6"
kaggle = "1.5.16"
Mako = "1.2.4"
MarkupSafe = "2.1.3"
mutagen = "1.47.0"
numpy = ">=1.26.0,<2.0"
outcome = "1.3.0.post0"
packaging = "24.1"
pandas = "2.1.4"
psycopg2-binary = "2.9.9"
pycryptodomex = "3.20.0"
pydantic = "2.5.3"
pydantic_core = "2.14.6"
PySocks = "1.7.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.0"
python-slugify = "8.0.4"
pytz = "2024.2"
requests = "2.32.2"
selenium = "4.23.1"
setuptools = "75.3.0"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "<2.0"
SQLAlchemy = "2.0.18"
sqlparse = "0.4.4"
text-unidecode = "1.3"
tqdm = "4.66.2"
trio = "0.26.1"
trio-websocket = "0.11.1"
typing_extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.1"
webdriver-manager = "4.0.2"
webencodings = "0.5.1"
websocket-client = "1.8.0"
websockets = "13.0.1"
wheel = "0.45.0"
wsproto = "1.2.0"
youtube-dl = "2021.12.17"
yt-dlp = "2024.3.10"
gunicorn = "21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

